workflows/react_distributed/actor/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3>=1.40.0
httpx
langchain
langgraph
langchain-openai
python-dotenv
langgraph-checkpoint-aws
psutil
